nanoops: move fp32→bf16 weight cast into fused_mlp_block wrapper
Mirrors nanchat's `Linear.forward` pattern (`self.weight.to(dtype=x.dtype)`):
callers don't need to remember to cast master weights down to activation
dtype before passing them in. The conditional `.to(x.dtype)` runs only
when dtypes differ (no-op if matching, preserving existing tests where
caller passes already-cast weights).

The `.to()` op is autograd-traced, so dW flowing back from
FusedMLPBlock.backward gets routed through `.to()`'s backward (which
upcasts dW_bf16 → dW_fp32) and accumulates onto the original fp32
weight — same flow as `Linear` already produces.

integration.py: drop the now-redundant `.to(x.dtype)` casts in
`_patched_block_forward`, pass `self.mlp.c_fc.weight` / `self.mlp.c_proj.weight`
directly.

Tests: existing parity tests (8) + nanchat integration tests with both
flags (2) all pass.

+++ nanoops/integration.py
@@ -149,11 +149,13 @@ def _patched_block_forward(self, x, ve, cos_sin, window_size, kv_cache):
     # `.contiguous()` is cheap when already contiguous (returns same tensor);
     # if attn produced a non-contiguous view (e.g. via stride tricks) the
     # fused kernel needs contiguous input for its index arithmetic.
+    # Weight dtype cast (fp32 master → bf16 activation) is handled inside
+    # `fused_mlp_block` itself, so we pass the raw module weights here.
     x_2d = x.reshape(B * T, C).contiguous()
     y_2d = _fused_mlp_block(
         x_2d, None,
-        self.mlp.c_fc.weight.to(x.dtype),
-        self.mlp.c_proj.weight.to(x.dtype),
+        self.mlp.c_fc.weight,
+        self.mlp.c_proj.weight,
     )
     return y_2d.reshape(B, T, C)
